+++ feature/main/build.gradle.kts
@@ -22,4 +22,5 @@ dependencies {
 
     // others
     implementation(libs.androidx.core.splashscreen)
+    implementation(libs.accompanist.systemuicontroller)
 }
